backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
numpy
scikit-learn
joblib
